% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.add.from.template}
\alias{beast.add.from.template}
\title{Copy XML elements from template}
\usage{
beast.add.from.template(bxml, btemplate, select.after.elt = NA,
  select.to.elt = NA, verbose = 1)
}

