% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.add.monophylyStatistic.for.taxonsets}
\alias{beast.add.monophylyStatistic.for.taxonsets}
\title{Add monophylyStatistics for each phylogenetic cluster to XML file}
\usage{
beast.add.monophylyStatistic.for.taxonsets(bxml, treeModel.id,
  taxonset.prefix = "c")
}

