% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beast.add.startingtree}
\alias{beast.add.startingtree}
\title{Add starting tree to XML}
\usage{
beast.add.startingtree(bxml, start.tree.newick,
  beast.newickid = "startingTree", beast.usingDates = "true",
  beast.brlunits = "years", verbose = 1)
}
\description{
Add starting tree to XML
}

