% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.add.tmrcaStatistic.for.taxonsets}
\alias{beast.add.tmrcaStatistic.for.taxonsets}
\title{Add tmrcaStatistic for phylogenetic clusters to XML file}
\usage{
beast.add.tmrcaStatistic.for.taxonsets(bxml, treeModel.id = "treeModel",
  taxonset.prefix = "c", includeStem = "false", verbose = 1)
}
\author{
Oliver Ratmann
}

