% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.adjust.gmrfSkyrideLikelihood}
\alias{beast.adjust.gmrfSkyrideLikelihood}
\title{Adjust dimension parameter of gmrfSkyrideLikelihood}
\usage{
beast.adjust.gmrfSkyrideLikelihood(bxml, df, verbose = 1)
}

