% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.adjust.prior.rootheight}
\alias{beast.adjust.prior.rootheight}
\title{Ajust lower bound of uniform prior on root height}
\usage{
beast.adjust.prior.rootheight(bxml, df,
  rootHeight.idref = "treeModel.rootHeight", verbose = 1)
}

