% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beast.choose.seq.by.clusters}
\alias{beast.choose.seq.by.clusters}
\title{Pick sequence data set by type of phylogenetic cluster}
\usage{
beast.choose.seq.by.clusters(df.seq, select, verbose = 1)
}
\description{
Pick sequence data set by type of phylogenetic cluster
}

