% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.set.SEQ}
\alias{beast.set.SEQ}
\title{Set SEQ field from sequences in sequences in \code{DNAbin} matrix format}
\usage{
beast.set.SEQ(df, seq, verbose = 1)
}
\value{
Taxon data.table with additional column SEQ that contains sequences as a character string.
}
\description{
Sequences whose name matches the TAXON_ID field are added to the taxon data.table.
}
\author{
Oliver Ratmann
}

