% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beast.set.TAXON_NAME}
\alias{beast.set.TAXON_NAME}
\title{Set TAXON_NAME field from columns in the taxon data.table.}
\usage{
beast.set.TAXON_NAME(df, select = c("TAXON_ID"), select.sep = "_")
}
\value{
Taxon data.table with additional column TAXON_NAME.
}
\description{
Set TAXON_NAME field from columns in the taxon data.table.
}

