% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2.R
\name{beast2.add.cluster.taxonsets}
\alias{beast2.add.cluster.taxonsets}
\title{No description yet}
\usage{
beast2.add.cluster.taxonsets(bxml, df, beast2.spec, verbose = 0)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

