% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2.R
\name{beast2.add.data}
\alias{beast2.add.data}
\title{No description yet}
\usage{
beast2.add.data(bxml, seq.PROT.RT, df, beast2.spec, verbose = 1)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

