% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2.R
\name{beast2.get.specifications}
\alias{beast2.get.specifications}
\title{No description yet}
\usage{
beast2.get.specifications(xml.dir = NA, xml.filename = NA,
  mcmc.length = 2e+07, bdsky.intervalNumber = 4, alignment.filter = NA,
  tip.log.stem = FALSE, cluster.log = FALSE, cluster.monophyletic = FALSE)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

