% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2.R
\name{beast2.get.xml}
\alias{beast2.get.xml}
\title{No description yet}
\usage{
beast2.get.xml(bxml.template, seq.PROT.RT, df, beast2.spec, ph = NULL,
  verbose = 1)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

