% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2output.R
\name{beast2out.combine.clu.trees}
\alias{beast2out.combine.clu.trees}
\title{No description yet}
\usage{
beast2out.combine.clu.trees(indir, file.info, beastlabel.idx.clu = 1,
  beastlabel.idx.hivn = 2, beastlabel.idx.hivd = 3,
  beastlabel.idx.hivs = 4, beastlabel.idx.samplecode = 6,
  beastlabel.idx.rate = NA, method.nodectime = "any", verbose = FALSE)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

