% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2output.R
\name{beast2out.extract.cluster.trees}
\alias{beast2out.extract.cluster.trees}
\title{No description yet}
\usage{
beast2out.extract.cluster.trees(mph, mph.info)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

