% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2output.R
\name{beast2out.get.log.evolrate}
\alias{beast2out.get.log.evolrate}
\title{No description yet}
\usage{
beast2out.get.log.evolrate(files, opt.burnin = opt.burnin)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

