% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2output.R
\name{beast2out.plot.cluster.trees}
\alias{beast2out.plot.cluster.trees}
\title{No description yet}
\usage{
beast2out.plot.cluster.trees(df.all, df.immu, df.viro, df.treatment, ph,
  ph.root.ctime, ph.tip.ctime, ph.prob = NA, df.node.ctime = NULL,
  df.rates = NULL, df.tips = NULL, end.ctime = 2013.3,
  cex.nodelabel = 0.5, cex.tiplabel = 0.5, file = NULL, pdf.width = 7,
  pdf.height = 20, pdf.xlim = NULL, label.select = c("cluster",
  "PatientA", "Trm", "isAcute", "RegionHospital"))
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

