% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2output.R
\name{beast2out.read.nexus.and.stats}
\alias{beast2out.read.nexus.and.stats}
\title{No description yet}
\usage{
beast2out.read.nexus.and.stats(file, tree.id = NA,
  method.node.stat = "any.node")
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

