% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2output.R
\name{beast2out.read.nodeidtree}
\alias{beast2out.read.nodeidtree}
\title{Private funtion to read tree such that tip.label and nodel.label include
information on the index on when a tip/node occurs in bstr}
\usage{
beast2out.read.nodeidtree(bstr, method.node.stat = "any.node")
}
\description{
Private funtion to read tree such that tip.label and nodel.label include
information on the index on when a tip/node occurs in bstr
}
\author{
Oliver Ratmann
}

