% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2output.R
\name{beast2out.tip.date.check}
\alias{beast2out.tip.date.check}
\title{No description yet}
\usage{
beast2out.tip.date.check(ph, fun, ...)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

