% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.beast.R
\name{beastxml.multilocus.hky.fixed.topology}
\alias{beastxml.multilocus.hky.fixed.topology}
\title{Generate multilocus XML file with HKY model and fixed starting tree}
\usage{
beastxml.multilocus.hky.fixed.topology(file.name, seq.select, phd,
  verbose = 1)
}
\description{
See Examples.
}
\author{
Oliver Ratmann
}

