% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_act_cpp}
\alias{calc_act_cpp}
\title{Does magic}
\usage{
calc_act_cpp(sample, sample_interval)
}
\arguments{
\item{sample}{sample}

\item{sample_interval}{sample interval}
}
\value{
the ACT
}
\description{
Does magic
}
