% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_esses.R
\name{calc_esses}
\alias{calc_esses}
\title{Calculates the Effective Sample Sizes}
\usage{
calc_esses(traces, sample_interval)
}
\arguments{
\item{traces}{a dataframe with traces with removed burn-in}

\item{sample_interval}{the interval in timesteps between samples}
}
\value{
the effective sample sizes
}
\description{
Calculates the Effective Sample Sizes
}
\author{
Richel Bilderbeek
}

