% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_state.R
\name{extract_operators_lines}
\alias{extract_operators_lines}
\title{Extract the lines out of a .xml.state file with the operators}
\usage{
extract_operators_lines(filename)
}
\description{
Extract the lines out of a .xml.state file with the operators
}
\author{
Richel Bilderbeek
}
