% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_posterior.R
\name{parse_beast_posterior}
\alias{parse_beast_posterior}
\title{Parses BEAST2 output files to a Cer2016 posterior}
\usage{
parse_beast_posterior(trees_filename, log_filename)
}
\arguments{
\item{trees_filename}{name of the BEAST2 .trees output file}

\item{log_filename}{name of the BEAST2 .trees output file}
}
\value{
a Cer2016 posterior
}
\description{
Parses BEAST2 output files to a Cer2016 posterior
}
\examples{
  trees_filename <- system.file(
   "extdata", "beast2_example_output.trees", package = "RBeast"
  )
  log_filename <- system.file(
   "extdata", "beast2_example_output.log", package = "RBeast"
  )
  posterior <- parse_beast_posterior(
    trees_filename = trees_filename,
    log_filename = log_filename
  )
  testit::assert(is_posterior(posterior))
}
\author{
Richel Bilderbeek
}

