% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beast2_trees.R
\name{read_beast2_trees}
\alias{read_beast2_trees}
\title{Extract a list of phylogenies from a BEAST2 posterior file}
\usage{
read_beast2_trees(filename)
}
\arguments{
\item{filename}{name of the BEAST2 posterior filename, usually ends with '.trees'}
}
\value{
a list of phylogenies of type 'phylo'
}
\description{
Extract a list of phylogenies from a BEAST2 posterior file
}
\examples{
  trees_file <- system.file(
    "extdata", "read_beast2_trees_example.trees", package = "RBeast"
  )
  testit::assert(file.exists(trees_file))
  posterior <- read_beast2_trees(trees_file)
  testit::assert(length(posterior) == 11)
  testit::assert(class(posterior[[1]]) == "phylo")
}
\author{
Richel J.C. Bilderbeek
}
