% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBeast2Trees.R
\name{readBeast2Trees}
\alias{readBeast2Trees}
\title{Extract a list of phylogenies from a BEAST2 posterior file}
\usage{
readBeast2Trees(file, opt.burnin = 0)
}
\arguments{
\item{file}{name of the BEAST2 posterior filename, usually ends with '.trees'}

\item{opt.burnin}{how many phylogenies to discard, a value of zero will keep all trees}
}
\value{
a list of phylogenies of type 'phylo'
}
\description{
Extract a list of phylogenies from a BEAST2 posterior file
}
\examples{
  trees_file <- "../inst/extdata/readBeast2TreesExample.trees"
  testit::assert(file.exists(trees_file))
  posterior <- readBeast2Trees(trees_file)
  testit::assert(length(posterior) == 10)
  testit::assert(class(posterior[[1]]) == "phylo")
}
\author{
Oliver Ratmann
}

