% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_burn_ins.R
\name{remove_burn_ins}
\alias{remove_burn_ins}
\title{Removed the burn-ins from a data frame}
\usage{
remove_burn_ins(traces, burn_in_fraction)
}
\arguments{
\item{traces}{a data frame with traces}

\item{burn_in_fraction}{the fraction that needs to be removed, must be [0,1>}
}
\value{
the data frame with the burn-in removed
}
\description{
Removed the burn-ins from a data frame
}
\author{
Richel J.C. Bilderbeek
}
