% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackedAreas.R
\name{runBeast}
\alias{runBeast}
\title{runBeast}
\usage{
runBeast(commandLine = "")
}
\arguments{
\item{commandLine}{Command-line string to pass to BEAST}
}
\description{
\code{runBeast} spins up BEAST
}
\details{
This function executes BEAST through rJava.
}
\examples{
if (1 == 2) {
  runBeast()
}
}

