% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.get.clusterprob}
\alias{treeannotator.get.clusterprob}
\title{No description yet}
\usage{
treeannotator.get.clusterprob(ph.beast, beastlabel.idx.clu = 1,
  beastlabel.idx.samplecode = 5, verbose = 1)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

