% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.get.edgewidth}
\alias{treeannotator.get.edgewidth}
\title{No description yet}
\usage{
treeannotator.get.edgewidth(ph, rates.df, scale.edgewidth = 12)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

