% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.get.immu.timeline}
\alias{treeannotator.get.immu.timeline}
\title{No description yet}
\usage{
treeannotator.get.immu.timeline(ph, df, df.immu, youngest.tip.ctime,
  end.ctime = 2013.3)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

