% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.get.phy}
\alias{treeannotator.get.phy}
\title{No description yet}
\usage{
treeannotator.get.phy(ph.beast, beastlabel.idx.clu = 1,
  beastlabel.idx.hivs = 4, beastlabel.idx.samplecode = 5,
  beastlabel.idx.rate = 6, verbose = 1, debug = 0)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

