% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.get.rates}
\alias{treeannotator.get.rates}
\title{No description yet}
\usage{
treeannotator.get.rates(ph, tip.df, nodelabel.idx.edgewidth = 5)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

