% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.get.tmrcas}
\alias{treeannotator.get.tmrcas}
\title{No description yet}
\usage{
treeannotator.get.tmrcas(ph.beast, beastlabel.idx.hivs = 4)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

