% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.get.viro.timeline}
\alias{treeannotator.get.viro.timeline}
\title{return lRNA measurements for treatment periods (going to be different
colors) for tips for which at least one RNA value is available.
PosRNA is translated relative to youngest tip calendar time}
\usage{
treeannotator.get.viro.timeline(ph, df, df.viro, youngest.tip.ctime,
  df.treatment = NULL, end.ctime = 2013.3)
}
\description{
return lRNA measurements for treatment periods (going to be different
colors) for tips for which at least one RNA value is available.
PosRNA is translated relative to youngest tip calendar time
}
\author{
Oliver Ratmann
}

