% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.nodelabels.getnodeheightHPD}
\alias{treeannotator.nodelabels.getnodeheightHPD}
\title{No description yet}
\usage{
treeannotator.nodelabels.getnodeheightHPD(ph, youngest.tip.ctime,
  node.label.hpd.l = 3, node.label.hpd.u = 4)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

