% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.ctimeline}
\alias{treeannotator.plot.ctimeline}
\title{No description yet}
\usage{
treeannotator.plot.ctimeline(ph, youngest.tip.ctime, end.ctime,
  col.bg = c(my.fade.col("black", 0.15), my.fade.col("black", 0.05)),
  col.txt = c(my.fade.col("black", 1), "transparent"), cex.txt = 0.5,
  add.yinch = 0.5)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

