% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.hpdbars}
\alias{treeannotator.plot.hpdbars}
\title{No description yet}
\usage{
treeannotator.plot.hpdbars(ph, ph.nodeheighthpds, col = "grey75", lwd = 4)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

