% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.immu.timeline}
\alias{treeannotator.plot.immu.timeline}
\title{No description yet}
\usage{
treeannotator.plot.immu.timeline(ph, ph.immu.timeline, immu.min = 150,
  immu.max = 800, immu.legend = c(200, 350, 500, immu.max),
  width.yinch = 0.15, add.yinch = -0.005, col.bg = cols[3],
  col.legend = cols[4], cex.txt = 0.2, lines.lwd = 0.2)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

