% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.node.ctime}
\alias{treeannotator.plot.node.ctime}
\title{No description yet}
\usage{
treeannotator.plot.node.ctime(df.node.ctime, ph.root.ctime, width.yinch = 0.1,
  add.yinch = 0.001, col.bg = "black", density = 30, lwd = 0.5)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

