% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.rates}
\alias{treeannotator.plot.rates}
\title{No description yet}
\usage{
treeannotator.plot.rates(ph, edge.width, add.xinch = -0.1, cex.rate = 0.5)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

