% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot}
\alias{treeannotator.plot}
\title{No description yet}
\usage{
treeannotator.plot(ph, ph.root.ctime, youngest.tip.ctime, df.all, df.viro,
  df.immu, df.treatment = NULL, df.tstem = NULL, df.rates = NULL,
  end.ctime = 2013.3, cex.nodelabel = 0.5, cex.tiplabel = 0.5,
  file = NULL, pdf.width = 7, pdf.height = 20)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

