% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.seronodeheightrange}
\alias{treeannotator.plot.seronodeheightrange}
\title{No description yet}
\usage{
treeannotator.plot.seronodeheightrange(ph, ph.seronodeheight,
  add.yinch = -0.05, width.yinch = 0.1, width.yinch.past.AnyPos_T1 = 0.02,
  col = "red")
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

