% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.tipmarks}
\alias{treeannotator.plot.tipmarks}
\title{No description yet}
\usage{
treeannotator.plot.tipmarks(ph, df.tips, add.xinch = 0, add.yinch = 0)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

