% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.tipstem.timeline}
\alias{treeannotator.plot.tipstem.timeline}
\title{No description yet}
\usage{
treeannotator.plot.tipstem.timeline(ph, youngest.tip.ctime, df.tstem,
  width.yinch = 0.15, add.yinch = 0, col.bg = "grey75", density = 30,
  lwd = 0.5)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

