% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.plot.viro.timeline}
\alias{treeannotator.plot.viro.timeline}
\title{No description yet}
\usage{
treeannotator.plot.viro.timeline(ph, ph.viro.timeline, viro.min = log10(300),
  viro.legend = c(3, 4, 5, 6), width.yinch = 0.2, add.yinch = 0.005,
  col.bg = "red", col.legend = "red", cex.txt = 0.2, lines.lwd = 0.2)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

