% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.read}
\alias{treeannotator.read}
\title{Read a BEAST treeannotator file. The node.label is set to a data.table
that contains the SIMMAP annotation for the interior nodes in
the newick tree.}
\usage{
treeannotator.read(file, add.to.tiplabel = NA, rate.multiplier = NA,
  round.digit = NA, verbose = 1)
}
\description{
Read a BEAST treeannotator file. The node.label is set to a data.table
that contains the SIMMAP annotation for the interior nodes in
the newick tree.
}
\author{
Oliver Ratmann
}

