% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.sero.getnodeheight.range}
\alias{treeannotator.sero.getnodeheight.range}
\title{No description yet}
\usage{
treeannotator.sero.getnodeheight.range(ph, df, youngest.tip.ctime)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

