% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.treeannotator.R
\name{treeannotator.tiplabel2df}
\alias{treeannotator.tiplabel2df}
\title{No description yet}
\usage{
treeannotator.tiplabel2df(x, beastlabel.idx.clu = 1,
  beastlabel.idx.hivn = 2, beastlabel.idx.hivd = 3,
  beastlabel.idx.hivs = 4, beastlabel.idx.samplecode = 5,
  beastlabel.idx.rate = 6)
}
\description{
No description yet
}
\author{
Oliver Ratmann
}

