% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beast.add.monophylyStatistic.for.taxonsets}
\alias{beast.add.monophylyStatistic.for.taxonsets}
\title{Add monophylyStatistics for each phylogenetic cluster to XML file}
\usage{
beast.add.monophylyStatistic.for.taxonsets(bxml, treeModel.id,
  taxonset.prefix = "c")
}
\description{
Add monophylyStatistics for each phylogenetic cluster to XML file
}

