% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.add.startingtree}
\alias{beast.add.startingtree}
\title{Add starting tree to XML}
\usage{
beast.add.startingtree(bxml, start.tree.newick,
  beast.newickid = "startingTree", beast.usingDates = "true",
  beast.brlunits = "years", verbose = 1)
}

