% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beast.add.tmrcaStatistic.for.taxonsets}
\alias{beast.add.tmrcaStatistic.for.taxonsets}
\title{Add tmrcaStatistic for phylogenetic clusters to XML file}
\usage{
beast.add.tmrcaStatistic.for.taxonsets(bxml, treeModel.id = "treeModel",
  taxonset.prefix = "c", includeStem = "false", verbose = 1)
}
\description{
Add tmrcaStatistic for phylogenetic clusters to XML file
}

