% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beast.adjust.gmrfSkyrideLikelihood}
\alias{beast.adjust.gmrfSkyrideLikelihood}
\title{Adjust dimension parameter of gmrfSkyrideLikelihood}
\usage{
beast.adjust.gmrfSkyrideLikelihood(bxml, df, verbose = 1)
}
\description{
Adjust dimension parameter of gmrfSkyrideLikelihood
}

