% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beast.adjust.prior.rootheight}
\alias{beast.adjust.prior.rootheight}
\title{Ajust lower bound of uniform prior on root height}
\usage{
beast.adjust.prior.rootheight(bxml, df,
  rootHeight.idref = "treeModel.rootHeight", verbose = 1)
}
\description{
Ajust lower bound of uniform prior on root height
}

