% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.choose.seq.by.clusters}
\alias{beast.choose.seq.by.clusters}
\title{Pick sequence data set by type of phylogenetic cluster}
\usage{
beast.choose.seq.by.clusters(df.seq, select, verbose = 1)
}
\author{
Oliver Ratmann
}

