% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast.R
\name{beast.set.TAXON_NAME}
\alias{beast.set.TAXON_NAME}
\title{Set TAXON_NAME field from columns in the taxon data.table.}
\usage{
beast.set.TAXON_NAME(df, select = c("TAXON_ID"), select.sep = "_")
}
\arguments{
\item{df}{Undocumented}

\item{select}{Undocumented}

\item{select.sep}{Undocumented}
}
\value{
Taxon data.table with additional column TAXON_NAME.
}
\author{
Oliver Ratmann
}

