% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.beast2output.R
\name{beast2out.read.trees}
\alias{beast2out.read.trees}
\title{Extract a list of phylogenies from a BEAST2 posterior file}
\usage{
beast2out.read.trees(file, opt.rescale.edge.length = 1, opt.burnin = 0)
}
\arguments{
\item{file}{name of the BEAST2 posterior filename, usually ends with '.trees'}

\item{opt.rescale.edge.length}{Rescaling factor of the phylogenies,
a value of 1.0 denotes that the original edge lengths are maintained}

\item{opt.burnin}{how many phylogenies to discard, an value of will keep all trees}
}
\value{
a list of phylogenies of type 'phylo'
}
\description{
Extract a list of phylogenies from a BEAST2 posterior file
}
\examples{
  trees_file <- "vignettes/example.trees"
  testit::assert(file.exists(trees_file))
  posterior <- beast2out.read.trees(trees_file)
  testit::assert(length(posterior) == 10)
  testit::assert(class(posterior[[1]]) == "phylo")
}
\author{
Oliver Ratmann
}

