% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_state.R
\name{extract_operators_lines}
\alias{extract_operators_lines}
\title{Extract the JSON lines out of a .xml.state file with the operators}
\usage{
extract_operators_lines(filename)
}
\arguments{
\item{filename}{name of the BEAST2 .xml.state output file}
}
\description{
Extract the JSON lines out of a .xml.state file with the operators
}
\author{
Richel Bilderbeek
}
