% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_log.R
\name{parse_beast_log}
\alias{parse_beast_log}
\title{Parses a BEAST2 .log output file}
\usage{
parse_beast_log(filename)
}
\arguments{
\item{filename}{name of the BEAST2 .log output file}
}
\value{
data frame with all the parameter estimates
}
\description{
Parses a BEAST2 .log output file
}
\examples{
  log_filename <- system.file(
    "extdata", "beast2_example_output.log", package = "RBeast"
  )
  estimates <- parse_beast_log(filename = log_filename)
  expected_names <- c(
    "Sample", "posterior", "likelihood",
    "prior", "treeLikelihood", "TreeHeight",
    "BirthDeath", "birthRate2", "relativeDeathRate2"
  )
  testit::assert(names(estimates) == expected_names)
}
\author{
Richel Bilderbeek
}

