% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_posterior.R
\name{parse_beast_posterior}
\alias{parse_beast_posterior}
\title{Parses BEAST2 output files to a Cer2016 posterior}
\usage{
parse_beast_posterior(trees_filename, log_filename)
}
\arguments{
\item{trees_filename}{name of the BEAST2 .trees output file}

\item{log_filename}{name of the BEAST2 .trees output file}
}
\value{
a Cer2016 posterior
}
\description{
Parses BEAST2 output files to a Cer2016 posterior
}
\examples{
  posterior <- parse_beast_posterior(
    trees_filename = find_path("beast2_example_output.trees"),
    log_filename = find_path("beast2_example_output.log")
  )
  testit::assert(is_posterior(posterior))
}
\author{
Richel Bilderbeek
}

